+++ spring-cloud-gateway-core/src/test/java/org/springframework/cloud/gateway/test/CustomBlockHoundIntegration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2019 the original author or authors.
+ * Copyright 2013-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package reactor.blockhound.integration;
+package org.springframework.cloud.gateway.test;
 
 import reactor.blockhound.BlockHound;
+import reactor.blockhound.integration.BlockHoundIntegration;
 
 /**
  * @author Tim Ysewyn
