+++ spring-cloud-gateway-core/src/test/resources/META-INF/services/reactor.blockhound.integration.BlockHoundIntegration
@@ -1,1 +1,1 @@
-reactor.blockhound.integration.CustomBlockHoundIntegration
+org.springframework.cloud.gateway.test.CustomBlockHoundIntegration
